Created CivicThemeMigrationManager service, updated validator to pass back array of messages.
